learn/voice-agent/requirements.txt
websockets
sounddevice
numpy
python-dotenv